Moved Helper component to separate file
